backend/pyproject.toml
[tool.poetry]
name = "cdb"
version = "1.208.2"
description = ""
authors = ["LionelB <lionel@lumographe.fr>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
fastapi = "^0.85.0"
uvicorn = "^0.18.0"
python-dotenv = "^0.21.0"
asyncpg = "^0.26.0"
sentry-sdk = "^1.5.11"
typer = "^0.6.0"
colorama = "^0.4.4"
shellingham = "^1.4.0"
dask = "^2022.4.0"
numpy = "^1.22.3"
pandas = "^1.4.2"
python-multipart = "^0.0.5"
PyJWT = "^2.4.0"
panda = "^0.3.1"
xlrd = "^2.0.1"
openpyxl = "^3.0.10"
StrEnum = "^0.4.7"
python-magic = "^0.4.26"
pydantic = {extras = ["email"], version = "^1.9.1"}
fastapi-asyncpg = "^1.0.1"
python-dateutil = "^2.8.2"
Jinja2 = "^3.1.2"
chardet = "^5.0.0"
httpx = "^0.23.0"
respx = "^0.20.0"
luhn-validator = "^1.1.4"
phonenumbers = "^8.12.56"
gql = {extras = ["aiohttp"], version = "^3.4.0"}
structlog = "^22.3.0"
structlog-sentry = "^2.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.1"
pytest-asyncio = "0.20.3"
pytest-mock = "^3.7.0"
pytest-testmon = "^1.3.7"
pytest-watch = "^4.2.0"
syrupy = "^3.0.5"

[tool.ruff]
select = [
    "E",  # pycodestyle
    "F",  # Pyflakes
    "B",  # flake8-bugbear
    "I",  # Isort
]
extend-exclude = ["cdb/api/_gen/schema_gql.py"]
ignore = [
    # Do not perform function call `XXX` in argument defaults.
    # FastAPI Depends() relies on this.
    "B008",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
